requirements.txt
fastapi
uvicorn
numpy
sympy
google-generativeai
google-ai-generativelanguage
python-dotenv